extract audioContext
